Integrate Morango sync resumption, keeping SoUD sync session alive between syncs
Fix progress handler

Keep sync session alive

Remove venv accidentally added during pre-commit fixes
