+++ kolibri/core/auth/management/utils.py
@@ -2,8 +2,11 @@
 Utility methods for syncing.
 """
 import getpass
+import json
 import logging
+import math
 import time
+from contextlib import contextmanager
 from functools import wraps
 
 import requests
@@ -12,13 +15,19 @@ from django.db.models.signals import post_delete
 from django.urls import reverse
 from django.utils.six.moves import input
 from morango.models import Certificate
-from morango.models import ScopeDefinition
+from morango.models import InstanceIDModel
+from morango.sync.controller import MorangoProfileController
 from six.moves.urllib.parse import urljoin
 
 from kolibri.core.auth.backends import FACILITY_CREDENTIAL_KEY
+from kolibri.core.auth.constants.morango_sync import DATA_PORTAL_SYNCING_BASE_URL
+from kolibri.core.auth.constants.morango_sync import PROFILE_FACILITY_DATA
 from kolibri.core.auth.constants.morango_sync import ScopeDefinitions
+from kolibri.core.auth.constants.morango_sync import State
+from kolibri.core.auth.models import dataset_cache
 from kolibri.core.auth.models import Facility
 from kolibri.core.auth.models import FacilityUser
+from kolibri.core.auth.sync_event_hook_utils import register_sync_event_handlers
 from kolibri.core.device.models import DevicePermissions
 from kolibri.core.device.utils import device_provisioned
 from kolibri.core.device.utils import provision_device
@@ -26,6 +35,10 @@ from kolibri.core.device.utils import set_device_settings
 from kolibri.core.discovery.utils.network.client import NetworkClient
 from kolibri.core.discovery.utils.network.errors import NetworkLocationNotFound
 from kolibri.core.discovery.utils.network.errors import URLParseError
+from kolibri.core.logger.utils.data import bytes_for_humans
+from kolibri.core.tasks.exceptions import UserCancelledError
+from kolibri.core.tasks.management.commands.base import AsyncCommand
+from kolibri.core.utils.lock import db_lock
 
 
 logger = logging.getLogger(__name__)
@@ -148,7 +161,16 @@ def get_dataset_id(baseurl, identifier=None, noninteractive=False):
         )
 
 
+def is_portal_sync(baseurl):
+    return baseurl == DATA_PORTAL_SYNCING_BASE_URL
+
+
 def get_baseurl(baseurl):
+    # if url matches data portal, no need to validate it
+    if is_portal_sync(baseurl):
+        return baseurl
+
+    # validate base url
     try:
         return NetworkClient(address=baseurl).base_url
     except URLParseError:
@@ -159,6 +181,22 @@ def get_baseurl(baseurl):
         )
     except NetworkLocationNotFound:
         raise CommandError("Unable to connect to: {}".format(baseurl))
+
+
+def get_network_connection(baseurl):
+    controller = MorangoProfileController(PROFILE_FACILITY_DATA)
+    network_connection = controller.create_network_connection(get_baseurl(baseurl))
+
+    # validate instance IDs are differemt, which would mean this device is trying to sync with itself
+    if (
+        InstanceIDModel.get_or_create_current_instance()[0].id
+        == network_connection.server_info["instance_id"]
+    ):
+        raise CommandError(
+            "Device can not sync with itself. Please recheck base URL and try again."
+        )
+
+    return network_connection
 
 
 def get_client_and_server_certs(
@@ -312,12 +350,17 @@ def provision_single_user_device(user_id):
     )
 
 
-def get_single_user_sync_filter(dataset_id, user_id, is_read):
-    scopedef = ScopeDefinition.objects.get(id=ScopeDefinitions.SINGLE_USER)
-    scope = scopedef.get_scope({"dataset_id": dataset_id, "user_id": user_id})
-    if is_read:
-        return str(scope.read_filter)
-    return str(scope.write_filter)
+def get_sync_filter(client_cert):
+    """
+    :type client_cert: Certificate
+    :return: Filter
+    """
+    scope = client_cert.get_scope()
+
+    if client_cert.scope_definition_id == ScopeDefinitions.SINGLE_USER:
+        return scope.read_filter
+
+    return scope.write_filter
 
 
 def run_once(f):
@@ -419,3 +462,319 @@ class GroupDeletion(object):
                 time.sleep(sleep)
 
         return total_count, all_deletions
+
+
+class MorangoSyncCommand(AsyncCommand):
+    """
+    Common methods for Morango sync commands
+    """
+
+    TRANSFER_MESSAGE = "{records_transferred}/{records_total}, {transfer_total}"
+
+    def _sync(self, sync_session_client, **options):  # noqa: C901
+        (username, no_push, no_pull, noninteractive, no_provision, keep_alive,) = (
+            options["username"],
+            options["no_push"],
+            options["no_pull"],
+            options["noninteractive"],
+            options["no_provision"],
+            options["keep_alive"],
+        )
+
+        client_cert = sync_session_client.sync_session.client_certificate
+        register_sync_event_handlers(sync_session_client.controller)
+        sync_filter = get_sync_filter(client_cert)
+
+        scope_params = json.loads(client_cert.scope_params)
+        dataset_id = scope_params.get("dataset_id")
+        user_id = scope_params.get("user_id", None)
+
+        dataset_cache.clear()
+        dataset_cache.activate()
+
+        if not noninteractive:
+            # output session ID for CLI user
+            logger.info("Session ID: {}".format(sync_session_client.sync_session.id))
+        logger.info("Syncing has been initiated (this may take a while)...")
+
+        try:
+            # pull from server
+            if not no_pull:
+                self._pull(
+                    sync_session_client,
+                    noninteractive,
+                    sync_filter,
+                )
+                # and push our own data to server
+            if not no_push:
+                self._push(
+                    sync_session_client,
+                    noninteractive,
+                    sync_filter,
+                )
+
+            if not no_provision:
+                with self._lock():
+                    if user_id:
+                        provision_single_user_device(user_id)
+                    else:
+                        create_superuser_and_provision_device(
+                            username, dataset_id, noninteractive=noninteractive
+                        )
+
+        except UserCancelledError:
+            if self.job:
+                self.job.extra_metadata.update(sync_state=State.CANCELLED)
+                self.job.save_meta()
+            logger.info("Syncing has been cancelled.")
+            return
+
+        if not keep_alive:
+            sync_session_client.sync_connection.close()
+
+        if self.job:
+            self.job.extra_metadata.update(sync_state=State.COMPLETED)
+            self.job.save_meta()
+
+        dataset_cache.deactivate()
+        logger.info("Syncing has been completed.")
+
+    @contextmanager
+    def _lock(self):
+        cancellable = False
+        # job can't be cancelled while locked
+        if self.job:
+            cancellable = self.job.cancellable
+            self.job.save_as_cancellable(cancellable=False)
+
+        with db_lock():
+            yield
+
+        if self.job:
+            self.job.save_as_cancellable(cancellable=cancellable)
+
+    def _raise_cancel(self, *args, **kwargs):
+        if self.is_cancelled() and (not self.job or self.job.cancellable):
+            raise UserCancelledError()
+
+    def _pull(
+        self,
+        sync_session_client,
+        noninteractive,
+        sync_filter,
+    ):
+        """
+        :type sync_session_client: morango.sync.syncsession.SyncSessionClient
+        :type noninteractive: bool
+        :type sync_filter: Filter
+        """
+        sync_client = sync_session_client.get_pull_client()
+        sync_client.signals.queuing.connect(self._raise_cancel)
+        sync_client.signals.transferring.connect(self._raise_cancel)
+
+        self._queueing_tracker_adapter(
+            sync_client.signals.queuing,
+            "Remotely preparing data",
+            State.REMOTE_QUEUING,
+            noninteractive,
+        )
+        self._transfer_tracker_adapter(
+            sync_client.signals.transferring,
+            "Receiving data ({})".format(self.TRANSFER_MESSAGE),
+            State.PULLING,
+            noninteractive,
+        )
+        self._queueing_tracker_adapter(
+            sync_client.signals.dequeuing,
+            "Locally integrating received data",
+            State.LOCAL_DEQUEUING,
+            noninteractive,
+        )
+
+        self._session_tracker_adapter(
+            sync_client.signals.session,
+            "Creating pull transfer session",
+            "Completed pull transfer session",
+        )
+
+        sync_client.initialize(sync_filter)
+
+        sync_client.run()
+        with self._lock():
+            sync_client.finalize()
+
+    def _push(
+        self,
+        sync_session_client,
+        noninteractive,
+        sync_filter,
+    ):
+        """
+        :type sync_session_client: morango.sync.syncsession.SyncSessionClient
+        :type noninteractive: bool
+        :type sync_filter: Filter
+        """
+        sync_client = sync_session_client.get_push_client()
+        sync_client.signals.transferring.connect(self._raise_cancel)
+
+        self._queueing_tracker_adapter(
+            sync_client.signals.queuing,
+            "Locally preparing data to send",
+            State.LOCAL_QUEUING,
+            noninteractive,
+        )
+        self._transfer_tracker_adapter(
+            sync_client.signals.transferring,
+            "Sending data ({})".format(self.TRANSFER_MESSAGE),
+            State.PUSHING,
+            noninteractive,
+        )
+        self._queueing_tracker_adapter(
+            sync_client.signals.dequeuing,
+            "Remotely integrating data",
+            State.REMOTE_DEQUEUING,
+            noninteractive,
+        )
+
+        self._session_tracker_adapter(
+            sync_client.signals.session,
+            "Creating push transfer session",
+            "Completed push transfer session",
+        )
+
+        with self._lock():
+            sync_client.initialize(sync_filter)
+
+        sync_client.run()
+
+        # we can't cancel remotely integrating data
+        if self.job:
+            self.job.save_as_cancellable(cancellable=False)
+
+        # allow server timeout since remotely integrating data can take a while and the request
+        # could timeout. In that case, we'll assume everything is good.
+        sync_client.finalize()
+
+    def _update_all_progress(self, progress_fraction, progress):
+        """
+        Override parent progress update callback to report from the progress tracker we're sent
+        """
+        if self.job:
+            self.job.update_progress(progress_fraction, 1.0)
+            self.job.extra_metadata.update(progress.extra_data)
+            self.job.save_meta()
+
+    def _session_tracker_adapter(self, signal_group, started_msg, completed_msg):
+        """
+        Attaches a signal handler to session creation signals
+
+        :type signal_group: morango.sync.syncsession.SyncSignalGroup
+        :type started_msg: str
+        :type completed_msg: str
+        """
+
+        @run_once
+        def session_creation(transfer_session):
+            """
+            A session is created individually for pushing and pulling
+            """
+            logger.info(started_msg)
+            if self.job:
+                self.job.extra_metadata.update(sync_state=State.SESSION_CREATION)
+
+        @run_once
+        def session_destruction(transfer_session):
+            if transfer_session.records_total == 0:
+                logger.info("There are no records to transfer")
+            logger.info(completed_msg)
+
+        signal_group.started.connect(session_creation)
+        signal_group.completed.connect(session_destruction)
+
+    def _transfer_tracker_adapter(
+        self, signal_group, message, sync_state, noninteractive
+    ):
+        """
+        Attaches a signal handler to pushing/pulling signals
+
+        :type signal_group: morango.sync.syncsession.SyncSignalGroup
+        :type message: str
+        :type sync_state: str
+        :type noninteractive: bool
+        """
+        tracker = self.start_progress(total=100)
+
+        def stats_msg(transfer_session):
+            transfer_total = (
+                transfer_session.bytes_sent + transfer_session.bytes_received
+            )
+            return message.format(
+                records_transferred=transfer_session.records_transferred,
+                records_total=transfer_session.records_total,
+                transfer_total=bytes_for_humans(transfer_total),
+            )
+
+        def stats(transfer_session):
+            logger.info(stats_msg(transfer_session))
+
+        def handler(transfer_session):
+            """
+            :type transfer_session: morango.models.core.TransferSession
+            """
+            if transfer_session.records_total > 0:
+                progress = (
+                    100
+                    * transfer_session.records_transferred
+                    / float(transfer_session.records_total)
+                )
+            else:
+                progress = 100
+
+            tracker.update_progress(
+                increment=math.ceil(progress - tracker.progress),
+                message=stats_msg(transfer_session),
+                extra_data=dict(
+                    bytes_sent=transfer_session.bytes_sent,
+                    bytes_received=transfer_session.bytes_received,
+                    sync_state=sync_state,
+                ),
+            )
+
+        if noninteractive or tracker.progressbar is None:
+            signal_group.started.connect(stats)
+            signal_group.in_progress.connect(stats)
+
+        signal_group.connect(handler)
+
+        # log one more time at end to capture in logging output
+        signal_group.completed.connect(stats)
+
+    def _queueing_tracker_adapter(
+        self, signal_group, message, sync_state, noninteractive
+    ):
+        """
+        Attaches a signal handler to queuing/dequeuing signals
+
+        :type signal_group: morango.sync.syncsession.SyncSignalGroup
+        :type message: str
+        :type sync_state: str
+        :type noninteractive: bool
+        """
+        tracker = self.start_progress(total=2)
+
+        def started(transfer_session):
+            dataset_cache.clear()
+            if noninteractive or tracker.progressbar is None:
+                logger.info(message)
+
+        def handler(transfer_session):
+            tracker.update_progress(
+                message=message, extra_data=dict(sync_state=sync_state)
+            )
+
+        if noninteractive or tracker.progressbar is None:
+            signal_group.started.connect(started)
+
+        signal_group.started.connect(started)
+        signal_group.started.connect(handler)
+        signal_group.completed.connect(handler)
